Update main.rs
Pin checks are removed for now, as the logic is hard and they're probably premature optimization at this point (if faster at all). I added an alternative pin-checking function that returns the attacking square for convenience later if they get added back. Legality checks now generally directly construct the follow-up position and determine whether the non-moving side is in check. This corrected the depth 4 perft from startpos.

Code for detecting whether a square is threatened by a pawn now looks for pawns in the correct place rather than the opposite direction. Perft totals are now correct to depth 5 (depth 4 totals for positions one ply into the game might be off, but the errors cancel if they exist). 

A number of commented-out chunks of code were removed.