Fix non-capture promotions
Turns out I really just put promotion code in the pawn captures section and not in the pawn pushing section of pawn pseudolegal movegen. That problem should be fixed now. 

Depth four Kiwipete perft is still 4085659 instead of 4085603, so there's a bit more debugging to be done. 

Something in the back of my mind tells me I copied too much code and forgot to change a variable name somewhere. Really hoping that's a false alarm.