Perft now correct to depth 4
Each non-960 position at https://analog-hors.github.io/webperft/ now gives the correct perft total up to depth 4 (in ply). The fen parser can't handle 960 at the moment, but the movegen is intended to be 960-compatible. Startpos is correct to depth 5. 

Unfortunately, the code is too slow to comfortably and quickly do higher depths for now. That's a definite area for improvement.