New naive eval and simple move ordering heuristic
Extremely basic evaluation that literally just sums up a material difference in centipawns (bishop = 3.1 pawns). The move ordering heuristic is actually more complicated than the evaluation, but the parameters for it are completely made up (by me). It is based on MVV-LVA but the structure of the move representation forces me to consider castling, en passant, and promotions separately. I decided (with no evidence or much consideration whatsoever) that castling was just slightly preferable to pawn takes pawn, that the null move is between not capturing anything and capturing a pawn with a king, and that promotions would receive an arbitrarily scaled bonus over standard moves based on the promotion choice, pulled a bit towards an average promotion value. I might change that later to keep the try queen -> try knight -> try rook -> try bishop order up.