Adding in a scuffed AB search
Cladonia now supports one of the minimal requirements for being vaguely okay. It is now possible to search to depth 8 in (quite a bit) less time than depth 6 used to take. The search is not end-of-game aware and may return strange results when mated positions show up. It is in desparate need of quiescence search and probably static exchange evaluation- horizon effect from startpos is pretty bad. Will probably remove the extremely bad "MVV-LVA pruning" thing soon that kept things to a max. branching factor of 12. Might also want to add a "number of available moves" option to the hce? Maybe not. Next big goal is probably UCI support so I can actually test some stuff. Also, the recent code is getting pretty ugly because it's a draft for before the abstraction is figured out. A rewrite is somewhere in Cladonia's future.